using provider, and cleaner code
