+++ example/lib/models/bluetoothPrinter.dart
@@ -1,0 +1,24 @@
+import 'package:flutter_pos_printer_platform/flutter_pos_printer_platform.dart';
+
+class BluetoothPrinter {
+  int? id;
+  String? deviceName;
+  String? address;
+  String? port;
+  String? vendorId;
+  String? productId;
+  bool? isBle;
+
+  PrinterType typePrinter;
+  bool? state;
+
+  BluetoothPrinter(
+      {this.deviceName,
+      this.address,
+      this.port,
+      this.state,
+      this.vendorId,
+      this.productId,
+      this.typePrinter = PrinterType.bluetooth,
+      this.isBle = false});
+}
