+++ example/lib/providers/thermal_print.dart
@@ -1,0 +1,331 @@
+import 'dart:async';
+import 'dart:convert';
+import 'dart:io';
+
+import 'package:flutter/foundation.dart';
+import 'package:flutter_pos_printer_platform/esc_pos_utils_platform/src/capability_profile.dart';
+import 'package:image/image.dart' as img;
+
+import 'package:flutter/services.dart';
+import 'package:flutter_pos_printer_platform/esc_pos_utils_platform/src/enums.dart';
+import 'package:flutter_pos_printer_platform/esc_pos_utils_platform/src/generator.dart';
+import 'package:flutter_pos_printer_platform/esc_pos_utils_platform/src/pos_column.dart';
+import 'package:flutter_pos_printer_platform/esc_pos_utils_platform/src/pos_styles.dart';
+import 'package:flutter_pos_printer_platform/flutter_pos_printer_platform.dart';
+
+import '../custom/image_utils.dart';
+import '../models/bluetoothPrinter.dart';
+
+class PrinterService extends ChangeNotifier {
+  PrinterService() {
+    if (Platform.isWindows) {
+      printerType = PrinterType.usb;
+    }
+    scan();
+    // listen change status of bluetooth connection
+    PrinterManager.instance.stateBluetooth.listen((status) {
+      print(' ----------------- status bt $status ------------------ ');
+      if (status == BTStatus.connected) {
+        isConnected = true;
+        notifyListeners();
+      }
+
+      if (status == BTStatus.connected && pendingTask != null) {
+        if (Platform.isAndroid) {
+          Future.delayed(const Duration(milliseconds: 1000), () {
+            PrinterManager.instance
+                .send(type: PrinterType.bluetooth, bytes: pendingTask!);
+            pendingTask = null;
+          });
+        } else if (Platform.isIOS) {
+          PrinterManager.instance
+              .send(type: PrinterType.bluetooth, bytes: pendingTask!);
+          pendingTask = null;
+        }
+      }
+    });
+    //  PrinterManager.instance.stateUSB is only supports on Android
+    PrinterManager.instance.stateUSB.listen((status) {
+      print(' ----------------- status usb $status ------------------ ');
+      if (Platform.isAndroid) {
+        if (status == USBStatus.connected && pendingTask != null) {
+          Future.delayed(const Duration(milliseconds: 1000), () {
+            PrinterManager.instance
+                .send(type: PrinterType.usb, bytes: pendingTask!);
+            pendingTask = null;
+          });
+        }
+      }
+    });
+  }
+  var printerType = PrinterType.bluetooth;
+
+  var isBle = false;
+  var reconnect = false;
+  bool get isReconnect => reconnect;
+  var isConnected = false;
+  var printerManager = PrinterManager.instance;
+  PrinterManager get printer => printerManager;
+
+  var devices = [];
+
+  final BTStatus _currentStatus = BTStatus.none;
+  // _currentUsbStatus is only supports on Android
+  // ignore: unused_field
+  final USBStatus _currentUsbStatus = USBStatus.none;
+  List<int>? pendingTask;
+  String _ipAddress = '';
+  String _port = '9100';
+  String get ipAddress => _ipAddress;
+  String get port => _port;
+  BluetoothPrinter? selectedPrinter;
+  // method to scan devices according PrinterType
+  void scan() {
+    devices.clear();
+    printerManager.discovery(type: printerType, isBle: isBle).listen((device) {
+      devices.add(BluetoothPrinter(
+        deviceName: device.name,
+        address: device.address,
+        isBle: isBle,
+        vendorId: device.vendorId,
+        productId: device.productId,
+        typePrinter: printerType,
+      ));
+      notifyListeners();
+    });
+  }
+
+  void disconnect() {
+    if (selectedPrinter != null) {
+      printerManager.disconnect(type: selectedPrinter!.typePrinter);
+    }
+    isConnected = false;
+    notifyListeners();
+  }
+
+  void setPort(String value) {
+    if (value.isEmpty) value = '9100';
+    _port = value;
+    var device = BluetoothPrinter(
+      deviceName: value,
+      address: _ipAddress,
+      port: _port,
+      typePrinter: PrinterType.network,
+      state: false,
+    );
+    selectDevice(device);
+  }
+
+  void setIpAddress(String value) {
+    _ipAddress = value;
+    var device = BluetoothPrinter(
+      deviceName: value,
+      address: _ipAddress,
+      port: _port,
+      typePrinter: PrinterType.network,
+      state: false,
+    );
+    selectDevice(device);
+  }
+
+  void changePrinterType(PrinterType type) {
+    printerType = type;
+    isBle = false;
+    initialState();
+    notifyListeners();
+    scan();
+  }
+
+  void changeReconnect(bool isReconnect) {
+    reconnect = isReconnect;
+    notifyListeners();
+  }
+
+  void changeIsBle(bool isB) {
+    isBle = isB;
+    initialState();
+    notifyListeners();
+    scan();
+  }
+
+  void initialState() {
+    selectedPrinter = null;
+    isConnected = false;
+  }
+
+  void selectDevice(BluetoothPrinter device) async {
+    if (selectedPrinter != null) {
+      if ((device.address != selectedPrinter!.address) ||
+          (device.typePrinter == PrinterType.usb &&
+              selectedPrinter!.vendorId != device.vendorId)) {
+        await PrinterManager.instance
+            .disconnect(type: selectedPrinter!.typePrinter);
+      }
+    }
+    selectedPrinter = device;
+    notifyListeners();
+    await connectDevice();
+  }
+
+  Future printReceiveTest() async {
+    List<int> bytes = [];
+
+    // Xprinter XP-N160I
+    final profile = await CapabilityProfile.load(name: 'XP-N160I');
+    // default profile
+    // final profile = await CapabilityProfile.load();
+
+    // PaperSize.mm80 or PaperSize.mm58
+    final generator = Generator(PaperSize.mm80, profile);
+    // bytes += generator.setGlobalCodeTable('CP1250');
+    bytes += generator.text('Test Print',
+        styles: const PosStyles(align: PosAlign.left));
+    bytes += generator.text('Product 1');
+    bytes += generator.text('Product 2');
+    // print accent
+    bytes += generator.text('Comunicación',
+        styles: const PosStyles(align: PosAlign.left, codeTable: 'CP1252'));
+
+    bytes += generator.emptyLines(1);
+
+    // sum width total column must be 12
+    bytes += generator.row([
+      PosColumn(
+          width: 8,
+          text: 'Lemon lime export quality per pound x 5 units',
+          styles: const PosStyles(align: PosAlign.left)),
+      PosColumn(
+          width: 4,
+          text: 'USD 2.00',
+          styles: const PosStyles(align: PosAlign.right)),
+    ]);
+
+    final ByteData data = await rootBundle.load('assets/ic_launcher.png');
+    if (data.lengthInBytes > 0) {
+      final Uint8List imageBytes = data.buffer.asUint8List();
+      // decode the bytes into an image
+      final decodedImage = img.decodeImage(imageBytes)!;
+      // Create a black bottom layer
+      // Resize the image to a 130x? thumbnail (maintaining the aspect ratio).
+      img.Image thumbnail = img.copyResize(decodedImage, height: 130);
+      // creates a copy of the original image with set dimensions
+      img.Image originalImg =
+          img.copyResize(decodedImage, width: 380, height: 130);
+      // fills the original image with a white background
+      img.fill(originalImg, color: img.ColorRgb8(255, 255, 255));
+      var padding = (originalImg.width - thumbnail.width) / 2;
+
+      //insert the image inside the frame and center it
+      drawImage(originalImg, thumbnail, dstX: padding.toInt());
+
+      // convert image to grayscale
+      var grayscaleImage = img.grayscale(originalImg);
+
+      bytes += generator.feed(1);
+      // bytes += generator.imageRaster(img.decodeImage(imageBytes)!, align: PosAlign.center);
+      bytes += generator.imageRaster(grayscaleImage, align: PosAlign.center);
+      bytes += generator.feed(1);
+    }
+
+    // PosCodeTable.westEur
+    bytes += generator.text('Special 1: àÀ èÈ éÉ ûÛ üÜ çÇ ôÔ',
+        styles: const PosStyles(codeTable: 'CP1252'));
+    bytes += generator.text('Special 2: blåbærgrød',
+        styles: const PosStyles(codeTable: 'CP1252'));
+    var esc = '\x1B';
+
+    // support arabic 22: arabic code page printer
+    bytes += Uint8List.fromList(List.from('${esc}t'.codeUnits)..add(22));
+    bytes += generator.textEncoded(Uint8List.fromList(utf8.encode('مرحبا بك')));
+
+    _printEscPos(bytes, generator);
+  }
+
+  /// print ticket
+  void _printEscPos(List<int> bytes, Generator generator) async {
+    var connectedTCP = false;
+    if (selectedPrinter == null) return;
+    var bluetoothPrinter = selectedPrinter!;
+
+    switch (bluetoothPrinter.typePrinter) {
+      case PrinterType.usb:
+        bytes += generator.feed(2);
+        bytes += generator.cut();
+        await printerManager.connect(
+            type: bluetoothPrinter.typePrinter,
+            model: UsbPrinterInput(
+                name: bluetoothPrinter.deviceName,
+                productId: bluetoothPrinter.productId,
+                vendorId: bluetoothPrinter.vendorId));
+        pendingTask = null;
+        break;
+      case PrinterType.bluetooth:
+        bytes += generator.cut();
+        await printerManager.connect(
+            type: bluetoothPrinter.typePrinter,
+            model: BluetoothPrinterInput(
+                name: bluetoothPrinter.deviceName,
+                address: bluetoothPrinter.address!,
+                isBle: bluetoothPrinter.isBle ?? false,
+                autoConnect: reconnect));
+        pendingTask = null;
+        if (Platform.isAndroid) pendingTask = bytes;
+        break;
+      case PrinterType.network:
+        bytes += generator.feed(2);
+        bytes += generator.cut();
+        connectedTCP = await printerManager.connect(
+            type: bluetoothPrinter.typePrinter,
+            model: TcpPrinterInput(ipAddress: bluetoothPrinter.address!));
+        if (!connectedTCP) print(' --- please review your connection ---');
+        break;
+      default:
+    }
+    if (bluetoothPrinter.typePrinter == PrinterType.bluetooth &&
+        Platform.isAndroid) {
+      if (_currentStatus == BTStatus.connected) {
+        printerManager.send(type: bluetoothPrinter.typePrinter, bytes: bytes);
+        pendingTask = null;
+      }
+    } else {
+      printerManager.send(type: bluetoothPrinter.typePrinter, bytes: bytes);
+      if (bluetoothPrinter.typePrinter == PrinterType.network) {
+        printerManager.disconnect(type: bluetoothPrinter.typePrinter);
+      }
+    }
+  }
+
+  // conectar dispositivo
+  Future connectDevice() async {
+    isConnected = false;
+    if (selectedPrinter == null) return;
+    switch (selectedPrinter!.typePrinter) {
+      case PrinterType.usb:
+        await printerManager.connect(
+            type: selectedPrinter!.typePrinter,
+            model: UsbPrinterInput(
+                name: selectedPrinter!.deviceName,
+                productId: selectedPrinter!.productId,
+                vendorId: selectedPrinter!.vendorId));
+        isConnected = true;
+        break;
+      case PrinterType.bluetooth:
+        await printerManager.connect(
+            type: selectedPrinter!.typePrinter,
+            model: BluetoothPrinterInput(
+                name: selectedPrinter!.deviceName,
+                address: selectedPrinter!.address!,
+                isBle: selectedPrinter!.isBle ?? false,
+                autoConnect: reconnect));
+        break;
+      case PrinterType.network:
+        await printerManager.connect(
+            type: selectedPrinter!.typePrinter,
+            model: TcpPrinterInput(ipAddress: selectedPrinter!.address!));
+        isConnected = true;
+        break;
+      default:
+    }
+    notifyListeners();
+  }
+}
